Working application, end of module 37

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -53,10 +53,14 @@ const Mutations = {
 	async deleteItem(parent, args, ctx, info) {
 		const where = { id: args.id };
 		// - Find Item 
-		const item = await ctx.db.query.item({ where }, `{ id title }`);
-
+		const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
 		// - Check user permissions 
-		// -> TODO 
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = 
+			ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+		if (!ownsItem || !hasPermissions) {
+			throw new Error('You don\'t have permission to do that');	
+		}
 		// - Delete Item
 		return ctx.db.mutation.deleteItem({ where }, info);
 	},
